Implement basic managed symbol resolution
Attach CLRMD DataTarget to the target process in passive mode,
and extract method information from the address. This only works for
processes that have the same bitness as the host, so we  will need to
extract the symbol resolution to a separate process or require that the
user run the target process under the same bitness as the profiler
process. Native symbol resolution is not implemented yet.